Adding stdint.h, utilizing uint8_t, uint16_t, and uint32_t explicitly
The reason for this, is that a 64 bit target architecture will interpret an "int" as 64 bits, and this code would generate PCIe transactions formatted as 64 bit requests. Some endpoint designs cannot handle a 64 bit request. Perhaps in the future this code can be enhanced to support length = 64, 32, 16, or 8. Currently it supports only three sizes.